requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.68.2
twilio==9.5.1
httpx==0.28.1
